+++ backend/book-service/src/main/java/kr/mybrary/bookservice/book/domain/BookWriteService.java
@@ -5,15 +5,15 @@ import kr.mybrary.bookservice.book.domain.dto.BookDtoMapper;
 import kr.mybrary.bookservice.book.domain.dto.request.BookCreateServiceRequest;
 import kr.mybrary.bookservice.book.domain.exception.BookAlreadyExistsException;
 import kr.mybrary.bookservice.book.persistence.Book;
-import kr.mybrary.bookservice.book.persistence.BookCategory;
-import kr.mybrary.bookservice.book.persistence.author.Author;
-import kr.mybrary.bookservice.book.persistence.author.BookAuthor;
+import kr.mybrary.bookservice.book.persistence.bookInfo.BookCategory;
+import kr.mybrary.bookservice.book.persistence.bookInfo.Author;
+import kr.mybrary.bookservice.book.persistence.bookInfo.BookAuthor;
 import kr.mybrary.bookservice.book.persistence.repository.AuthorRepository;
 import kr.mybrary.bookservice.book.persistence.repository.BookCategoryRepository;
 import kr.mybrary.bookservice.book.persistence.repository.BookRepository;
 import kr.mybrary.bookservice.book.persistence.repository.TranslatorRepository;
-import kr.mybrary.bookservice.book.persistence.translator.BookTranslator;
-import kr.mybrary.bookservice.book.persistence.translator.Translator;
+import kr.mybrary.bookservice.book.persistence.bookInfo.BookTranslator;
+import kr.mybrary.bookservice.book.persistence.bookInfo.Translator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
