+++ backend/book-service/src/main/java/kr/mybrary/bookservice/book/persistence/Book.java
@@ -10,8 +10,9 @@ import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
-import kr.mybrary.bookservice.book.persistence.author.BookAuthor;
-import kr.mybrary.bookservice.book.persistence.translator.BookTranslator;
+import kr.mybrary.bookservice.book.persistence.bookInfo.BookAuthor;
+import kr.mybrary.bookservice.book.persistence.bookInfo.BookCategory;
+import kr.mybrary.bookservice.book.persistence.bookInfo.BookTranslator;
 import kr.mybrary.bookservice.global.BaseEntity;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
