+++ backend/book-service/src/main/java/kr/mybrary/bookservice/book/persistence/bookInfo/BookTranslator.java
@@ -1,4 +1,4 @@
-package kr.mybrary.bookservice.book.persistence.translator;
+package kr.mybrary.bookservice.book.persistence.bookInfo;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
