+++ backend/book-service/src/main/java/kr/mybrary/bookservice/book/persistence/repository/BookCategoryRepository.java
@@ -1,7 +1,7 @@
 package kr.mybrary.bookservice.book.persistence.repository;
 
 import java.util.Optional;
-import kr.mybrary.bookservice.book.persistence.BookCategory;
+import kr.mybrary.bookservice.book.persistence.bookInfo.BookCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BookCategoryRepository extends JpaRepository<BookCategory, Long> {
