+++ backend/book-service/src/main/java/kr/mybrary/bookservice/book/persistence/repository/TranslatorRepository.java
@@ -1,7 +1,7 @@
 package kr.mybrary.bookservice.book.persistence.repository;
 
 import java.util.Optional;
-import kr.mybrary.bookservice.book.persistence.translator.Translator;
+import kr.mybrary.bookservice.book.persistence.bookInfo.Translator;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TranslatorRepository extends JpaRepository<Translator, Long> {
