+++ backend/book-service/src/test/java/kr/mybrary/bookservice/book/persistence/AuthorRepositoryTest.java
@@ -3,7 +3,7 @@ package kr.mybrary.bookservice.book.persistence;
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
 
-import kr.mybrary.bookservice.book.persistence.author.Author;
+import kr.mybrary.bookservice.book.persistence.bookInfo.Author;
 import kr.mybrary.bookservice.book.persistence.repository.AuthorRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
