+++ backend/book-service/src/test/java/kr/mybrary/bookservice/book/persistence/BookCategoryRepositoryTest.java
@@ -3,6 +3,7 @@ package kr.mybrary.bookservice.book.persistence;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
+import kr.mybrary.bookservice.book.persistence.bookInfo.BookCategory;
 import kr.mybrary.bookservice.book.persistence.repository.BookCategoryRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
