+++ backend/book-service/src/test/java/kr/mybrary/bookservice/book/persistence/TranslatorRepositoryTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import kr.mybrary.bookservice.book.persistence.repository.TranslatorRepository;
-import kr.mybrary.bookservice.book.persistence.translator.Translator;
+import kr.mybrary.bookservice.book.persistence.bookInfo.Translator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
